refactor: algolia-utilファイル名変更
