+++ functions/src/utils/algolia-util.ts
